requirements.txt
alembic==1.11.1
blinker==1.6.2
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.3.2
Flask-JWT-Extended==4.4.4
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
google-api-core==2.11.0
google-auth==2.18.1
google-cloud-core==2.3.2
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
mysql-connector-python==8.0.33
mysqlclient==2.1.1
protobuf==3.20.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyJWT==2.7.0
python-dotenv==1.0.0
python-magic==0.4.27
requests==2.31.0
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.15
typing_extensions==4.6.1
urllib3==1.26.16
Werkzeug==2.3.4
